Fixed Sign In and modified Get Database Data

+++ app/src/main/java/com/example/birdtrail_opsc7312/LandingPage.kt
@@ -43,9 +43,10 @@ class LandingPage : AppCompatActivity() {
             if (userID != null) {
 
 
+
                 for (user in GlobalClass.userData)
                 {
-                    if (user.userID == userID.toInt())
+                    if (user.userID == userID)
                     {
                         GlobalClass.currentUser = user
                     }
